Vaccine Study Sample Data

A sample dataset based on the simulations of a cholera vaccine trial in Heydrich-Perez et al. (2014) (doi:10.1111/biom.12184 ) except with 3000 individuals in 250 groups rather than 10000 in 500.

Format

a dataset with 6 variables and 3000 rows

Ythe outcome (0 - no cholera; 1 - cholera)

X1an individual's age (in decades)

X2an individual's distance from river

Aan indicator of vaccination (0 - no vaccine; 1 - vaccine)

Ban indicator of participation (0 - did not participant in vaccine trial, 1 - did participate)

groupgroup membership

References

Perez-Heydrich, C., Hudgens, M. G., Halloran, M. E., Clemens, J. D., Ali, M., & Emch, M. E. (2014). Assessing effects of cholera vaccination in the presence of interference. Biometrics, 70(3), 731-741.